Time: 5 ms (41.68%), Space: 56.7 MB (10.88%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,17 +1,17 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        double max =0.0;
-        int sum = 0;
+        double max = 0.0;
+        double sum = 0.0;
         for(int i = 0; i < k; i++){
             sum += nums[i];
             
         }
-        max = (double)sum / k;
+        max = sum / k;
 
         for(int i = k; i < nums.length; i++){
             sum += nums[i] - nums[i-k];
 
-            max = Math.max(max , (double)sum / k);
+            max = Math.max(max , sum / k);
         }
 
         return max;
